Update README.md
http://www.enkivillage.com/a-vs-an.html:
The “a” vs “an” rule applies to acronyms, not just complete words. If the first letter of an acronym is pronounced with a vowel sound, then it is preceded by “an.” Consonants with vowel sounds can include: f, h, l, m, n, r, s, and x. Below are three examples:

She bought an SUV